fix: timetable ui 컴포넌트분리
