+++ src/components/TimetablePage/DefaultPage/DefaultPage.module.scss
@@ -2,17 +2,6 @@
 
 .page {
   &__title {
-    text-align: left;
-    font-family: NanumSquare, serif;
-    font-size: 30px;
-    font-weight: 800;
-    color: #175c8e;
-    width: 100%;
-    height: 32px;
-    margin: 0 0 21px;
-    padding-bottom: 26px;
-    border-bottom: 2px solid #175c8e;
-
     &--sub {
       font-size: 20px;
       font-weight: 800;
@@ -21,13 +10,6 @@
       color: #175c8e;
       margin-bottom: 26px;
     }
-  }
-
-  &__content {
-    display: grid;
-    grid-template-columns: 766px 1fr;
-    grid-template-rows: 535px 336px;
-    gap: 66px 28px;
   }
 
   &__filter {
