+++ src/components/TimetablePage/DefaultPage/MyLectureTimetable.tsx
@@ -1,21 +1,30 @@
+/* eslint-disable react/require-default-props */
 /* eslint-disable no-restricted-imports */
 import ErrorBoundary from 'components/common/ErrorBoundary';
 import LoadingSpinner from 'components/common/LoadingSpinner';
 import React from 'react';
 import useImageDownload from 'utils/hooks/useImageDownload';
-import SemesterListbox from '../MyLectureTimetable/SemesterListbox';
-import CurrentSemesterTimetable from '../MyLectureTimetable/CurrentSemesterTable';
+import { TimetableDayLectureInfo } from 'interfaces/Lecture';
 import styles from './DefaultPage.module.scss';
+import SemesterListbox from '../MyLectureTimetable/SemesterListbox';
+import Timetable from '../MyLectureTimetable/Timetable';
 
-export default function MyLectureTimetable() {
+interface Props {
+  lectures: TimetableDayLectureInfo[][];
+  selectedLectureIndex?: number;
+  similarSelectedLecture?: TimetableDayLectureInfo[][];
+}
+
+export default function MyLectureTimetable(
+  { lectures, similarSelectedLecture, selectedLectureIndex }: Props,
+) {
   const { onImageDownload: onTimetableImageDownload, divRef: timetableRef } = useImageDownload();
+
   return (
     <div className={styles['page__timetable-wrap']}>
       <div className={styles.page__filter}>
         <div className={styles.page__semester}>
-          <React.Suspense fallback={<LoadingSpinner size="50" />}>
-            <SemesterListbox />
-          </React.Suspense>
+          <SemesterListbox />
         </div>
         <button
           type="button"
@@ -29,7 +38,15 @@ export default function MyLectureTimetable() {
       <div ref={timetableRef} className={styles.page__timetable}>
         <ErrorBoundary fallbackClassName="loading">
           <React.Suspense fallback={<LoadingSpinner size="50" />}>
-            <CurrentSemesterTimetable />
+            <Timetable
+              lectures={lectures}
+              similarSelectedLecture={similarSelectedLecture}
+              selectedLectureIndex={selectedLectureIndex}
+              columnWidth={55}
+              firstColumnWidth={52}
+              rowHeight={21}
+              totalHeight={456}
+            />
           </React.Suspense>
         </ErrorBoundary>
       </div>
